\name{DyaDA-internal}
\title{Dyadic Data Analysis (internal functions)}
\alias{DminS.comp}
\alias{ISI.comp}
\alias{ISI.info}
\alias{ordering.matrix}
\alias{prepare.block.matrix}
\alias{prepare.SRM.matrix}
\alias{prepmat}
\alias{print.isi}
\alias{print.hlineartest}
\alias{plot.sigdyads}
\alias{print.dyadr}
\alias{print.dyadicc}
\alias{print.rp}
\alias{print.dyadkappa}
\alias{print.dyadkappa2}
\alias{print.srmRREffects}
\alias{print.srmRRVars}
\alias{print.srmRRrelVars}
\alias{print.srmCovFTest}
\alias{print,srmWithintTest}
\alias{print.reciptest1}
\alias{print.reciptest2}
\alias{squaremat}
\alias{swaprowcol}
\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  DyaDA package. 
  For user information ask for help about \code{\link[DyaDA]{DyaDA}}.}
\keyword{internal}
