\name{getexpecR}

\alias{getexpecR}

\title{Mathematical Expectancy for Dietz's R statistic}

\description{
Function to compute mathematical expectancy for Dietz's R statistic.}

\usage{
   getexpecR(X,Y)
}

\arguments{
  \item{X}{Original matrix X.}
  \item{Y}{Original matrix Y.}
}

\details{
\code{getexpecR} computes mathematical expectancy for Dietz's R statistic.
}

\value{

\item{expecR}{Mathematical Expectancy for Dietz's R statistic.}
}

\references{

Dietz, E. J. (1983). Permutation tests for association between two distance matrices. \emph{Systematic Zoology}, \emph{32}, 21-26.
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu>, Han de Vries <J.deVries1@uu.nl>, & David A. Kenny <david.kenny@uconn.edu>.

}

\seealso{ \code{\link{getSER}}, \code{\link{getR}}, \code{\link{dietzRtest}}.
}

\keyword{misc}

\keyword{htest}


