\name{getpearson}

\alias{getpearson}

\title{Pearson's Correlation between two matrices}

\description{
Function to compute Pearson's statistic for two matrices.}

\usage{
   getpearson(X,Y)
}

\arguments{
  \item{X}{Original matrix X.}
  \item{Y}{Original matrix Y.}
}

\details{
\code{getpearson} computes Pearson's rho statistic for quantifying association between two matrices.
}

\value{

\item{pearson}{Pearson's correlation.}
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu>, Han de Vries <J.deVries1@uu.nl>, & David A. Kenny <david.kenny@uconn.edu>.

}

\keyword{misc}

\keyword{htest}

