\name{getspearman}

\alias{getspearman}

\title{Spearman's Correlation between two matrices}

\description{
Function to compute Spearman's statistic for two matrices.}

\usage{
   getspearman(X,Y)
}

\arguments{
  \item{X}{Original matrix X.}
  \item{Y}{Original matrix Y.}
}

\details{
\code{getspearman} computes Spearman's rho statistic for quantifying association between two matrices.
}

\value{

\item{spearman}{Spearman's correlation.}
}

\author{
David Leiva <dleivaur@ub.edu>, Antonio Solanas <antonio.solanas@ub.edu>, Han de Vries <J.deVries1@uu.nl>, & David A. Kenny <david.kenny@uconn.edu>.

}

\keyword{misc}

\keyword{htest}

